requirements.txt
google-generativeai>=0.7.2
gradio>=4.44.0
python-dotenv>=1.0.1
reportlab>=4.2.5
